src/instrument_detect/instrument_detector/requirements.txt
transformers
accelerate
bitsandbytes
safetensors
ray
ray[serve]
soundfile
jupyter
pillow
librosa
torchcodec
loguru
vllm>=0.8.0